Refactor: Update import paths for player separation functions and related utilities
- Changed import paths for `separatePlayers` and its variants to use the new location in `td-scenes/battle`.
- Introduced new utility functions `getMainTurnCount` and `isContinuousActiveTurn` to manage game state logic.
- Updated tests to reflect the new import paths for the above functions.
- Ensured deprecated functions are marked appropriately.
