+++ src/js/custom-battle-events/confrontation-two-braver/procedures/on-state-animation/shinya-battle-when-yuuya-hp-is-full.ts
@@ -1,6 +1,6 @@
 import { CustomStateAnimationProps } from "../../../../td-scenes/battle/custom-battle-event";
+import { separatePlayersFromCurrentState } from "../../../../td-scenes/battle/separate-players";
 import { ConditionalAnimation } from "../../../get-animation-if-conditional-met";
-import { separatePlayersFromCurrentState } from "../../../separate-players";
 import { shinyaAttackShoutWhenYuuyaHPIsFull } from "../../animation/shinya-attack-shout-when-yuuya-hp-is-full";
 import { ConfrontationTwoBraverProps } from "../../props";
