+++ src/js/custom-battle-events/confrontation-two-braver/procedures/on-state-animation/shinya-battle-when-yuuya-takes-damage.ts
@@ -1,6 +1,6 @@
 import { CustomStateAnimationProps } from "../../../../td-scenes/battle/custom-battle-event";
+import { separatePlayersFromCurrentState } from "../../../../td-scenes/battle/separate-players";
 import { ConditionalAnimation } from "../../../get-animation-if-conditional-met";
-import { separatePlayersFromCurrentState } from "../../../separate-players";
 import { shinyaAttackShoutWhenYuuyaTakesDamage } from "../../animation/shinya-attack-shout-when-yuuya-takes-damage";
 import { ConfrontationTwoBraverProps } from "../../props";
