+++ src/js/custom-battle-events/confrontation-two-braver/procedures/on-state-animation/shinya-pilot-skill.ts
@@ -1,10 +1,10 @@
 import { CustomStateAnimationProps } from "../../../../td-scenes/battle/custom-battle-event";
+import { separatePlayersFromCurrentState } from "../../../../td-scenes/battle/separate-players";
 import { ConditionalAnimation } from "../../../get-animation-if-conditional-met";
 import { isEnemyAdvantage } from "../../../is-enemy-advantage";
 import { isEvenMatch } from "../../../is-even-match";
 import { isPlayerPilotSkillActivatedFromCurrentState } from "../../../is-pilot-skill-activated";
 import { isPlayerAdvantage } from "../../../is-player-advantage";
-import { separatePlayersFromCurrentState } from "../../../separate-players";
 import { shinyaPilotSkillWhenEvenMatch } from "../../animation/shinya-pilot-skill-when-even-match";
 import { shinyaPilotSkillWhenShinyaHasAdvantage } from "../../animation/shinya-pilot-skill-when-shinya-has-advantage";
 import { shinyaPilotSkillWhenYuuyaHasAdvantage } from "../../animation/shinya-pilot-skill-when-yuuya-has-advantage";
