+++ src/js/custom-battle-events/confrontation-two-braver/procedures/on-state-animation/yuuya-finish-blow.ts
@@ -1,7 +1,7 @@
 import { CustomStateAnimationProps } from "../../../../td-scenes/battle/custom-battle-event";
+import { separatePlayersFromCurrentState } from "../../../../td-scenes/battle/separate-players";
 import { ConditionalAnimation } from "../../../get-animation-if-conditional-met";
 import { hasDeliveredFinishBlow } from "../../../has-delivered-finish-blow";
-import { separatePlayersFromCurrentState } from "../../../separate-players";
 import { yuuyaShoutWhenHeDeliversFinalBlow } from "../../animation/yuuya-shout-when-he-delivers-final-blow";
 import { ConfrontationTwoBraverProps } from "../../props";
