+++ src/js/custom-battle-events/pilot-skill-tutorial-01/procedures/on-state-animation/tsubasa-shout.ts
@@ -1,10 +1,10 @@
 import { CustomStateAnimationProps } from "../../../../td-scenes/battle/custom-battle-event";
+import { separatePlayersFromCurrentState } from "../../../../td-scenes/battle/separate-players";
 import { ConditionalAnimation } from "../../../get-animation-if-conditional-met";
 import { getPlayerBattleCount } from "../../../get-battle-count";
 import { hasDeliveredFinishBlow } from "../../../has-delivered-finish-blow";
 import { isEnemyBurstActivatedFromCurrentState } from "../../../is-burst-activated";
 import { isEnemyPilotSkillActivatedFromCurrentState } from "../../../is-pilot-skill-activated";
-import { separatePlayersFromCurrentState } from "../../../separate-players";
 import { tsubasaBurstShout } from "../../animation/tsubasa-burst-shout";
 import { tsubasaFinishBlowShout } from "../../animation/tsubasa-finish-blow-shout";
 import { tsubasaFirstAttackShout } from "../../animation/tsubasa-first-attack-shout";
