+++ src/js/custom-battle-events/queen-of-tragedy/procedure/on-state-animation/yuuya-attack-when-he-has-advantage.ts
@@ -1,8 +1,8 @@
 import { empty } from "../../../../animation/delay";
 import { CustomStateAnimationProps } from "../../../../td-scenes/battle/custom-battle-event";
+import { separatePlayersFromCurrentState } from "../../../../td-scenes/battle/separate-players";
 import { ConditionalAnimation } from "../../../get-animation-if-conditional-met";
 import { isPlayerAdvantage } from "../../../is-player-advantage";
-import { separatePlayersFromCurrentState } from "../../../separate-players";
 import { yuuyaAttackShoutWhenHeHasAdvantage } from "../../animation/yuuya-attack-shout-when-he-has-advantage";
 import { QueenOfTragedyProps } from "../../props";
