+++ src/js/custom-battle-events/queen-of-tragedy/procedure/on-state-animation/yuuya-attack-when-he-has-disadvantage.ts
@@ -1,8 +1,8 @@
 import { empty } from "../../../../animation/delay";
 import { CustomStateAnimationProps } from "../../../../td-scenes/battle/custom-battle-event";
+import { separatePlayersFromCurrentState } from "../../../../td-scenes/battle/separate-players";
 import { ConditionalAnimation } from "../../../get-animation-if-conditional-met";
 import { isEnemyAdvantage } from "../../../is-enemy-advantage";
-import { separatePlayersFromCurrentState } from "../../../separate-players";
 import { yuuyaAttackShoutWhenHeHasDisadvantage } from "../../animation/yuuya-attack-shout-when-he-has-disadvantage";
 import { QueenOfTragedyProps } from "../../props";
