+++ src/js/custom-battle-events/zero-defense-tutorial/procedures/before-last-state/execute-damage-race-if-needed.ts
@@ -1,5 +1,5 @@
 import { LastStateEventProps } from "../../../../td-scenes/battle/custom-battle-event";
-import { separatePlayers } from "../../../separate-players";
+import { separatePlayers } from "../../../../td-scenes/battle/separate-players";
 import { ZeroDefenseTutorialProps } from "../../props";
 import { ZeroDefenseTutorialState } from "../../state";
 import { damageRace } from "../../stories/damage-race";
