+++ src/js/td-scenes/battle/get-main-turn-count.ts
