+++ src/js/td-scenes/battle/is-continuous-active-turn.ts
