+++ test/js/td-scenes/battle/is-continuous-active-turn.test.ts
@@ -1,6 +1,6 @@
 import { EMPTY_GAME_STATE, GameState } from "gbraver-burst-core";
 
-import { isContinuousActiveTurn } from "../../../src/js/custom-battle-events/is-continuous-active-turn";
+import { isContinuousActiveTurn } from "../../../../src/js/td-scenes/battle/is-continuous-active-turn";
 
 /** TurnStartのゲームステート */
 const TURN_START: GameState = {
